+++ README.md
@@ -19,3 +19,17 @@
   * 5개 일치, 보너스 볼 일치 (30000000원) - 0개
   * 6개 일치 (2000000000원) - 0개
   * 수익률 표시 = 당첨금/구매금액
+  
+## 리팩토링
+* 패키지 분리하기
+* hashcode, equals 구현
+* 테스트 코드를 간결하게 작성가능하도록 구조 변경
+* input 값의 validate 추가 구현
+  * 6개값이 안들어왔을 때
+  * 45보다 큰 값이 들어왔을 때
+  * 중복된 값이 들어왔을 때
+* 상수값 제거 하기
+* 로또객체에서 get으로 꺼내지말고, 메시지를 보내는 방식으로 바꿔봐라. 
+  * ex) get해서 비교 X -> isMatch로 구현 / int matchCount = userLotto.match(winningLotto)
+* 테스트 케이스 추가
+  * 등수별 테스트
