+++ src/main/java/lotto/domain/LottoNumberGeneratorStrategy.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.List;
