+++ src/main/java/lotto/domain/LottoNumbers.java
@@ -1,9 +1,7 @@
-package lotto;
+package lotto.domain;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 public class LottoNumbers {
     private List<Integer> lottoNumbers;
