+++ src/main/java/lotto/domain/LottoNumbersList.java
@@ -1,8 +1,7 @@
-package lotto;
+package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.function.Consumer;
 
 public class LottoNumbersList  {
