+++ src/test/java/lotto/LastWonLottoNumberTest.java
@@ -1,12 +1,9 @@
 package lotto;
 
-import org.junit.jupiter.api.BeforeEach;
+import lotto.domain.LastWonLottoNumber;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-
-import java.util.ArrayList;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
