+++ src/test/java/lotto/LottoNumbersListTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.*;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
