+++ src/test/java/lotto/PurchaseLottoTest.java
@@ -1,8 +1,7 @@
 package lotto;
 
+import lotto.domain.PurchaseLotto;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 
 
 import static org.assertj.core.api.Assertions.assertThat;
